move getFieldAttribute into TestUtils
