+++ x-pack/plugin/ql/src/test/java/org/elasticsearch/xpack/ql/optimizer/OptimizerRulesTests.java
@@ -7,6 +7,7 @@
 package org.elasticsearch.xpack.ql.optimizer;
 
 import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.xpack.ql.TestUtils;
 import org.elasticsearch.xpack.ql.expression.Alias;
 import org.elasticsearch.xpack.ql.expression.Expression;
 import org.elasticsearch.xpack.ql.expression.Expressions;
@@ -60,7 +61,6 @@ import org.elasticsearch.xpack.ql.tree.NodeInfo;
 import org.elasticsearch.xpack.ql.tree.Source;
 import org.elasticsearch.xpack.ql.type.DataType;
 import org.elasticsearch.xpack.ql.type.DataTypes;
-import org.elasticsearch.xpack.ql.type.EsField;
 import org.elasticsearch.xpack.ql.util.StringUtils;
 
 import java.time.ZoneId;
@@ -69,7 +69,6 @@ import java.util.List;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
-import static java.util.Collections.emptyMap;
 import static java.util.Collections.singletonList;
 import static org.elasticsearch.xpack.ql.TestUtils.equalsOf;
 import static org.elasticsearch.xpack.ql.TestUtils.fieldAttribute;
@@ -157,15 +156,7 @@ public class OptimizerRulesTests extends ESTestCase {
     }
 
     private static FieldAttribute getFieldAttribute() {
-        return getFieldAttribute("a");
-    }
-
-    private static FieldAttribute getFieldAttribute(String name) {
-        return getFieldAttribute(name, INTEGER);
-    }
-
-    private static FieldAttribute getFieldAttribute(String name, DataType dataType) {
-        return new FieldAttribute(EMPTY, name, new EsField(name + "f", dataType, emptyMap(), true));
+        return TestUtils.getFieldAttribute("a");
     }
 
     //
@@ -548,9 +539,9 @@ public class OptimizerRulesTests extends ESTestCase {
 
     // 1 < a AND a < 3 AND 2 < b AND b < 4 AND c < 4 -> (1 < a < 3) AND (2 < b < 4) AND c < 4
     public void testCombineMultipleComparisonsIntoRange() {
-        FieldAttribute fa = getFieldAttribute("a");
-        FieldAttribute fb = getFieldAttribute("b");
-        FieldAttribute fc = getFieldAttribute("c");
+        FieldAttribute fa = TestUtils.getFieldAttribute("a");
+        FieldAttribute fb = TestUtils.getFieldAttribute("b");
+        FieldAttribute fc = TestUtils.getFieldAttribute("c");
 
         ZoneId zoneId = randomZone();
         GreaterThan agt1 = new GreaterThan(EMPTY, fa, ONE, zoneId);
@@ -1040,9 +1031,9 @@ public class OptimizerRulesTests extends ESTestCase {
 
     // (a = 1 AND b = 3 AND c = 4) OR (a = 2 AND b = 3 AND c = 4) -> (b = 3 AND c = 4) AND (a = 1 OR a = 2)
     public void testBooleanSimplificationCommonExpressionSubstraction() {
-        FieldAttribute fa = getFieldAttribute("a");
-        FieldAttribute fb = getFieldAttribute("b");
-        FieldAttribute fc = getFieldAttribute("c");
+        FieldAttribute fa = TestUtils.getFieldAttribute("a");
+        FieldAttribute fb = TestUtils.getFieldAttribute("b");
+        FieldAttribute fc = TestUtils.getFieldAttribute("c");
 
         Expression a1 = equalsOf(fa, ONE);
         Expression a2 = equalsOf(fa, TWO);
@@ -1414,7 +1405,7 @@ public class OptimizerRulesTests extends ESTestCase {
 
     // a == 1 AND a == 2 -> nop for date/time fields
     public void testPropagateEquals_ignoreDateTimeFields() {
-        FieldAttribute fa = getFieldAttribute("a", DataTypes.DATETIME);
+        FieldAttribute fa = TestUtils.getFieldAttribute("a", DataTypes.DATETIME);
         Equals eq1 = equalsOf(fa, ONE);
         Equals eq2 = equalsOf(fa, TWO);
         And and = new And(EMPTY, eq1, eq2);
@@ -1564,8 +1555,8 @@ public class OptimizerRulesTests extends ESTestCase {
     }
 
     public void testTwoEqualsDifferentFields() throws Exception {
-        FieldAttribute fieldOne = getFieldAttribute("ONE");
-        FieldAttribute fieldTwo = getFieldAttribute("TWO");
+        FieldAttribute fieldOne = TestUtils.getFieldAttribute("ONE");
+        FieldAttribute fieldTwo = TestUtils.getFieldAttribute("TWO");
 
         Or or = new Or(EMPTY, equalsOf(fieldOne, ONE), equalsOf(fieldTwo, TWO));
         Expression e = new CombineDisjunctionsToIn().rule(or);
@@ -1670,7 +1661,7 @@ public class OptimizerRulesTests extends ESTestCase {
         FieldAttribute fa = getFieldAttribute();
         IsNull isNull = new IsNull(EMPTY, fa);
 
-        And nestedAnd = new And(EMPTY, lessThanOf(getFieldAttribute("b"), ONE), lessThanOf(getFieldAttribute("c"), ONE));
+        And nestedAnd = new And(EMPTY, lessThanOf(TestUtils.getFieldAttribute("b"), ONE), lessThanOf(TestUtils.getFieldAttribute("c"), ONE));
         And and = new And(EMPTY, isNull, nestedAnd);
         And top = new And(EMPTY, and, lessThanOf(fa, ONE));
 
@@ -1689,7 +1680,7 @@ public class OptimizerRulesTests extends ESTestCase {
             greaterThanOf(new Div(EMPTY, new Add(EMPTY, fa, ONE), TWO), ONE),
             greaterThanOf(new Add(EMPTY, fa, TWO), ONE)
         );
-        Expression kept = new And(EMPTY, isNull, lessThanOf(getFieldAttribute("b"), THREE));
+        Expression kept = new And(EMPTY, isNull, lessThanOf(TestUtils.getFieldAttribute("b"), THREE));
         And and = new And(EMPTY, nullified, kept);
 
         Expression optimized = new PropagateNullable().rule(and);
@@ -1729,8 +1720,8 @@ public class OptimizerRulesTests extends ESTestCase {
 
     public void testCombineFilters() throws Exception {
         EsRelation relation = relation();
-        GreaterThan conditionA = greaterThanOf(getFieldAttribute("a"), ONE);
-        LessThan conditionB = lessThanOf(getFieldAttribute("b"), TWO);
+        GreaterThan conditionA = greaterThanOf(TestUtils.getFieldAttribute("a"), ONE);
+        LessThan conditionB = lessThanOf(TestUtils.getFieldAttribute("b"), TWO);
 
         Filter fa = new Filter(EMPTY, relation, conditionA);
         Filter fb = new Filter(EMPTY, fa, conditionB);
@@ -1740,11 +1731,11 @@ public class OptimizerRulesTests extends ESTestCase {
 
     public void testPushDownFilter() throws Exception {
         EsRelation relation = relation();
-        GreaterThan conditionA = greaterThanOf(getFieldAttribute("a"), ONE);
-        LessThan conditionB = lessThanOf(getFieldAttribute("b"), TWO);
+        GreaterThan conditionA = greaterThanOf(TestUtils.getFieldAttribute("a"), ONE);
+        LessThan conditionB = lessThanOf(TestUtils.getFieldAttribute("b"), TWO);
 
         Filter fa = new Filter(EMPTY, relation, conditionA);
-        List<FieldAttribute> projections = singletonList(getFieldAttribute("b"));
+        List<FieldAttribute> projections = singletonList(TestUtils.getFieldAttribute("b"));
         Project project = new Project(EMPTY, fa, projections);
         Filter fb = new Filter(EMPTY, project, conditionB);
 
@@ -1754,12 +1745,12 @@ public class OptimizerRulesTests extends ESTestCase {
 
     public void testPushDownFilterThroughAgg() throws Exception {
         EsRelation relation = relation();
-        GreaterThan conditionA = greaterThanOf(getFieldAttribute("a"), ONE);
-        LessThan conditionB = lessThanOf(getFieldAttribute("b"), TWO);
+        GreaterThan conditionA = greaterThanOf(TestUtils.getFieldAttribute("a"), ONE);
+        LessThan conditionB = lessThanOf(TestUtils.getFieldAttribute("b"), TWO);
         GreaterThanOrEqual aggregateCondition = greaterThanOrEqualOf(new Count(EMPTY, ONE, false), THREE);
 
         Filter fa = new Filter(EMPTY, relation, conditionA);
-        List<FieldAttribute> projections = singletonList(getFieldAttribute("b"));
+        List<FieldAttribute> projections = singletonList(TestUtils.getFieldAttribute("b"));
         // invalid aggregate but that's fine cause its properties are not used by this rule
         Aggregate aggregate = new Aggregate(EMPTY, fa, emptyList(), emptyList());
         Filter fb = new Filter(EMPTY, aggregate, new And(EMPTY, aggregateCondition, conditionB));
@@ -1773,7 +1764,7 @@ public class OptimizerRulesTests extends ESTestCase {
 
     public void testIsNotNullOnIsNullField() {
         EsRelation relation = relation();
-        var fieldA = getFieldAttribute("a");
+        var fieldA = TestUtils.getFieldAttribute("a");
         Expression inn = isNotNull(fieldA);
         Filter f = new Filter(EMPTY, relation, inn);
 
@@ -1782,7 +1773,7 @@ public class OptimizerRulesTests extends ESTestCase {
 
     public void testIsNotNullOnOperatorWithOneField() {
         EsRelation relation = relation();
-        var fieldA = getFieldAttribute("a");
+        var fieldA = TestUtils.getFieldAttribute("a");
         Expression inn = isNotNull(new Add(EMPTY, fieldA, ONE));
         Filter f = new Filter(EMPTY, relation, inn);
         Filter expected = new Filter(EMPTY, relation, new And(EMPTY, isNotNull(fieldA), inn));
@@ -1792,8 +1783,8 @@ public class OptimizerRulesTests extends ESTestCase {
 
     public void testIsNotNullOnOperatorWithTwoFields() {
         EsRelation relation = relation();
-        var fieldA = getFieldAttribute("a");
-        var fieldB = getFieldAttribute("b");
+        var fieldA = TestUtils.getFieldAttribute("a");
+        var fieldB = TestUtils.getFieldAttribute("b");
         Expression inn = isNotNull(new Add(EMPTY, fieldA, fieldB));
         Filter f = new Filter(EMPTY, relation, inn);
         Filter expected = new Filter(EMPTY, relation, new And(EMPTY, new And(EMPTY, isNotNull(fieldA), isNotNull(fieldB)), inn));
@@ -1803,7 +1794,7 @@ public class OptimizerRulesTests extends ESTestCase {
 
     public void testIsNotNullOnFunctionWithOneField() {
         EsRelation relation = relation();
-        var fieldA = getFieldAttribute("a");
+        var fieldA = TestUtils.getFieldAttribute("a");
         var pattern = L("abc");
         Expression inn = isNotNull(
             new And(EMPTY, new TestStartsWith(EMPTY, fieldA, pattern, false), greaterThanOf(new Add(EMPTY, ONE, TWO), THREE))
@@ -1817,8 +1808,8 @@ public class OptimizerRulesTests extends ESTestCase {
 
     public void testIsNotNullOnFunctionWithTwoFields() {
         EsRelation relation = relation();
-        var fieldA = getFieldAttribute("a");
-        var fieldB = getFieldAttribute("b");
+        var fieldA = TestUtils.getFieldAttribute("a");
+        var fieldB = TestUtils.getFieldAttribute("b");
         var pattern = L("abc");
         Expression inn = isNotNull(new TestStartsWith(EMPTY, fieldA, fieldB, false));
